display confirm delete dialog for intakes

+++ lib/ui/views/intakes/intakes_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:mona/data/model/medication_intake.dart';
 import 'package:mona/data/providers/medication_intake_provider.dart';
+import 'package:mona/ui/widgets/dialogs.dart';
 import 'package:mona/ui/widgets/main_page_wrapper.dart';
 import 'package:provider/provider.dart';
 
@@ -19,16 +20,22 @@ class IntakesPage extends StatelessWidget {
             itemBuilder: (context, index) {
               MedicationIntake intake =
                   medicationIntakeProvider.takenIntakes[index];
-              final intakeDate = intake.takenDateTime ?? intake.scheduledDateTime; // TODO make taken non nullable and remove scheduled
+              final intakeDate =
+                  intake.takenDateTime ?? intake.scheduledDateTime;
+              // TODO make taken non nullable and remove scheduled
               final dateText = DateFormat.yMMMd().format(intakeDate);
-              final sideText = intake.side?.label != null ? ' • ${intake.side!.label}' : '';
+              final sideText =
+                  intake.side?.label != null ? ' • ${intake.side!.label}' : '';
               return ListTile(
                 title: Text(dateText),
                 subtitle: Text('${intake.dose} mg$sideText'),
                 trailing: IconButton(
                   icon: const Icon(Icons.delete_outline),
-                  onPressed: () {
-                    medicationIntakeProvider.deleteIntake(intake);
+                  onPressed: () async {
+                    final confirmed = await Dialogs.confirmDelete(context);
+                    if (confirmed == true) {
+                      medicationIntakeProvider.deleteIntake(intake);
+                    }
                   },
                 ),
               );
